fix: 認証済みユーザー provider の責務を保護ルートに移す
トップページで未ログイン状態の currentUser を root provider に保持したまま、ログイン後の保護ページへ遷移して useCurrentUser が失敗する状態を避けるため。
