+++ src/app/(protected)/layout.tsx
@@ -1,12 +1,19 @@
+import { AuthenticatedUserProvider } from '@/app/_providers/currentUser';
 import { requireUser } from '@/lib/server/session';
 import type { ReactNode } from 'react';
 
-// ログイン必須ルートのゲート。未ログインなら requireUser がログインへ誘導する。
-// <html>/<body> と provider は共有 root layout（src/app/layout.tsx）が持つ。
-const ProtectedLayout = async ({ children }: { children: ReactNode }) => {
-  await requireUser();
+export const dynamic = 'force-dynamic';
 
-  return <>{children}</>;
+// ログイン必須ルートのゲート。未ログインなら requireUser がログインへ誘導する。
+// root layout は全体共通の provider だけを持ち、認証済みユーザーはここで提供する。
+const ProtectedLayout = async ({ children }: { children: ReactNode }) => {
+  const session = await requireUser();
+
+  return (
+    <AuthenticatedUserProvider currentUser={session.user}>
+      {children}
+    </AuthenticatedUserProvider>
+  );
 };
 
 export default ProtectedLayout;
